fix: move DebouncedCommandInput
